Error handling, minor refactoring

+++ scripts/step03_explore_examples/02_advanced_rag/src/ingest.py
@@ -2,9 +2,7 @@ import logging
 import re
 import sys
 
-from langchain_community.document_loaders import PyMuPDFLoader
-from langchain_community.document_loaders import GitLoader
-
+from langchain_community.document_loaders import GitLoader, PyMuPDFLoader
 from utils.rag import split_docs_into_chunks
 
 from helpers.factory import setup_components
@@ -14,21 +12,31 @@ log = logging.getLogger(__name__)
 
 
 def load_documents_from_github():
-    # Load the documents from a GitHub repository
-    loader = GitLoader(
-        clone_url="https://github.com/SAP-docs/btp-cloud-platform",
-        repo_path="./gen/btp-cloud-platform",
-        file_filter=lambda file_path: re.match(
-            r"^./gen/btp-cloud-platform/docs/10-concepts/.*.md$", file_path
-        ),
-        branch="main",
-    )
-    text_documents = loader.load()
-    log.info("Getting the documents from the GitHub repository ...")
+    """
+    Load documents from a GitHub repository using GitLoader and split them into chunks.
 
-    # Split the documents into chunks
-    chunks = split_docs_into_chunks(documents=text_documents)
-    return chunks
+    Returns:
+        List of document chunks.
+    """
+    log.info("Getting the documents from the GitHub repository...")
+    try:
+        # Load the documents from a GitHub repository using GitLoader
+        loader = GitLoader(
+            clone_url="https://github.com/SAP-docs/btp-cloud-platform",
+            repo_path="./gen/btp-cloud-platform",
+            file_filter=lambda file_path: re.match(
+                r"^./gen/btp-cloud-platform/docs/10-concepts/.*.md$", file_path
+            ),
+            branch="main",
+        )
+        text_documents = loader.load()
+
+        # Split the documents into chunks
+        chunks = split_docs_into_chunks(documents=text_documents)
+        return chunks
+    except Exception as e:
+        log.error(f"Error loading documents from GitHub: {str(e)}")
+        return []
 
 
 def ingest_podcasts():
@@ -37,30 +45,41 @@ def ingest_podcasts():
     """
     _, _, db = setup_components(PODCASTS_TABLE_NAME)
 
-    # Clear any existing documents
-    db.delete(filter={})
+    try:
+        # Clear any existing documents
+        log.info("Deleting existing podcast documents from the table...")
+        db.delete(filter={})
+        log.info("Existing podcast documents deleted successfully!")
 
-    # Podcast data
-    episodes = [
-        {
-            "url": "https://sap-podcast-bucket.s3.amazonaws.com/the-future-of-supply-chain/The_Future_of_Supply_Chain_Episode_64_transcript.pdf",
-            "title": "Future of Supply Chain: Episode 64: Proactively Planning for Risk in Your Supply Chain with Everstream's Koray Kose and SAP's Volker Wilhelm",
-            "episode": 64,
-        },
-        {
-            "url": "https://sap-podcast-bucket.s3.amazonaws.com/the-future-of-supply-chain/The_Future_of_Supply_Chain_Episode_65_transcript.pdf",
-            "title": "The Future of Supply Chain: Episode 65: Grounding Your Supply Chain in Data with Google Cloud’s Paula Natoli",
-            "episode": 65,
-        },
-    ]
+        # Podcast data
+        episodes = [
+            {
+                "url": "https://sap-podcast-bucket.s3.amazonaws.com/the-future-of-supply-chain/The_Future_of_Supply_Chain_Episode_64_transcript.pdf",
+                "title": "Future of Supply Chain: Episode 64: Proactively Planning for Risk in Your Supply Chain with Everstream's Koray Kose and SAP's Volker Wilhelm",
+                "episode": 64,
+            },
+            {
+                "url": "https://sap-podcast-bucket.s3.amazonaws.com/the-future-of-supply-chain/The_Future_of_Supply_Chain_Episode_65_transcript.pdf",
+                "title": "The Future of Supply Chain: Episode 65: Grounding Your Supply Chain in Data with Google Cloud’s Paula Natoli",
+                "episode": 65,
+            },
+        ]
 
-    # Process and add documents for each episode
-    for ep in episodes:
-        loader = PyMuPDFLoader(ep["url"])
-        documents = load_and_process_documents(loader, ep["title"], ep["episode"])
-        db.add_documents(documents)
+        # Process and add documents for each episode
+        for ep in episodes:
+            loader = PyMuPDFLoader(ep["url"])
+            documents = load_and_process_documents(loader, ep["title"], ep["episode"])
+            if documents:
+                db.add_documents(documents)
+                log.info(f"Added documents for episode {ep['episode']} - {ep['title']}")
+            else:
+                log.warning(
+                    f"No documents found for episode {ep['episode']} - {ep['title']}"
+                )
 
-    log.info("Retriever created successfully.")
+        log.info("Retriever created successfully.")
+    except Exception as e:
+        log.error(f"Error during podcast ingestion: {str(e)}")
 
 
 def load_and_process_documents(loader, podcast_title, episode):
@@ -87,19 +106,29 @@ def load_and_process_documents(loader, podcast_title, episode):
 
 
 def ingest_sap_docs():
+    """
+    Ingest SAP documentation into HANA DB.
+    """
     _, _, db = setup_components(SAP_DOCS_TABLE_NAME)
 
     chunks = load_documents_from_github()
 
-    # Delete already existing documents from the table
-    log.info("Deleting existing sap btp docs from the table ...")
-    db.delete(filter={})
-    log.success("SAP btp docs chunks deleted successfully!")
+    if not chunks:
+        log.warning("No document chunks to ingest.")
+        return
 
-    # add the loaded document chunks to the HANA DB
-    log.info("Adding SAP btp docs chunks to the HANA DB ...")
-    db.add_documents(chunks)
-    log.success("Added SAP btp docs successfully!")
+    try:
+        # Delete already existing documents from the table
+        log.info("Deleting existing SAP BTP documents from the table...")
+        db.delete(filter={})
+        log.info("Existing SAP BTP documents deleted successfully!")
+
+        # Add the loaded document chunks to the HANA DB
+        log.info("Adding SAP BTP document chunks to the HANA DB...")
+        db.add_documents(chunks)
+        log.success("Added SAP btp docs successfully!")
+    except Exception as e:
+        log.error(f"Error during SAP documents ingestion: {str(e)}")
 
 
 def main():
